Helper to call a function
